@@ -30,8 +30,8 @@ dependencies:
   cloud_firestore: 4.8.0
   cloud_firestore_platform_interface: 5.15.0
   cloud_firestore_web: 3.6.0
-  dropdown_button2: 2.0.0
-  file_picker: 5.2.6
+  dropdown_button2: 2.3.9
+  file_picker: 6.1.1
   firebase_auth: 4.6.2
   firebase_auth_platform_interface: 6.15.2
   firebase_auth_web: 5.5.2
@@ -45,21 +45,21 @@ dependencies:
   firebase_storage_platform_interface: 4.4.2
   firebase_storage_web: 3.6.2
   flutter_animate: 4.1.1+1
-  flutter_cache_manager: 3.3.0
-  flutter_plugin_android_lifecycle: 2.0.9
+  flutter_cache_manager: 3.3.1
+  flutter_plugin_android_lifecycle: 2.0.17
   flutter_slidable: 
     git:
       url: https://github.com/letsar/flutter_slidable.git
       ref: d10a912aae1fb0da330d9c83169f15ed2b0b42a2
-  font_awesome_flutter: 10.1.0
+  font_awesome_flutter: 10.6.0
   from_css_color: 2.0.0
   go_router: 7.1.1
   google_fonts: 4.0.3
-  google_sign_in: 6.0.2
-  google_sign_in_android: 6.1.8
-  google_sign_in_ios: 5.6.1
-  google_sign_in_platform_interface: 2.4.0
-  google_sign_in_web: 0.11.0+2
+  google_sign_in: 6.1.5
+  google_sign_in_android: 6.1.20
+  google_sign_in_ios: 5.6.4
+  google_sign_in_platform_interface: 2.4.2
+  google_sign_in_web: 0.12.1
   image_picker: 1.0.4
   image_picker_android: 0.8.8
   image_picker_for_web: 3.0.1
@@ -68,19 +68,19 @@ dependencies:
   intl: 0.18.1
   json_path: 0.6.2
   mime_type: 1.0.0
-  page_transition: 2.0.4
+  page_transition: 2.1.0
   path_provider: 2.0.14
   path_provider_android: 2.0.25
   path_provider_foundation: 2.2.2
   path_provider_platform_interface: 2.0.6
-  plugin_platform_interface: 2.1.3
-  provider: 6.0.4
+  plugin_platform_interface: 2.1.6
+  provider: 6.0.5
   rxdart: 0.27.7
-  shared_preferences: 2.0.15
-  shared_preferences_android: 2.1.0
-  shared_preferences_ios: 2.1.1
-  shared_preferences_platform_interface: 2.2.0
-  shared_preferences_web: 2.1.0
+  shared_preferences: 2.2.2
+  shared_preferences_android: 2.2.1
+  shared_preferences_foundation: 2.3.4
+  shared_preferences_platform_interface: 2.3.1
+  shared_preferences_web: 2.2.1
   sign_in_with_apple: 4.3.0
   sign_in_with_apple_platform_interface: 1.0.0
   sign_in_with_apple_web: 1.0.1
@@ -93,11 +93,11 @@ dependencies:
   url_launcher_android: 6.0.27
   url_launcher_ios: 6.1.4
   url_launcher_platform_interface: 2.1.2
-  video_player: 2.6.0
-  video_player_android: 2.4.3
-  video_player_avfoundation: 2.4.2
-  video_player_platform_interface: 6.1.0
-  video_player_web: 2.0.15
+  video_player: 2.8.1
+  video_player_android: 2.4.10
+  video_player_avfoundation: 2.5.1
+  video_player_platform_interface: 6.2.1
+  video_player_web: 2.1.2
 
 
   # The following adds the Cupertino Icons font to your application.
