@@ -2,7 +2,7 @@ import '/auth/firebase_auth/auth_util.dart';
 import '/backend/backend.dart';
 import '/componentes/bs_create_product/bs_create_product_widget.dart';
 import '/componentes/bs_edit_product/bs_edit_product_widget.dart';
-import '/componentes/bs_side_bar/bs_side_bar_widget.dart';
+import '/componentes/bs_side_bar_admin/bs_side_bar_admin_widget.dart';
 import '/flutter_flow/flutter_flow_animations.dart';
 import '/flutter_flow/flutter_flow_icon_button.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
@@ -122,7 +122,7 @@ class _ProductsHomeWidgetState extends State<ProductsHomeWidget>
                             : FocusScope.of(context).unfocus(),
                         child: Padding(
                           padding: MediaQuery.viewInsetsOf(context),
-                          child: BsSideBarWidget(),
+                          child: BsSideBarAdminWidget(),
                         ),
                       );
                     },
