@@ -1,23 +1,21 @@
 import '/auth/firebase_auth/auth_util.dart';
-import '/backend/backend.dart';
-import '/componentes/bs_side_bar/bs_side_bar_widget.dart';
-import '/flutter_flow/flutter_flow_animations.dart';
-import '/flutter_flow/flutter_flow_icon_button.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
 import '/flutter_flow/flutter_flow_widgets.dart';
-import 'productos_widget.dart' show ProductosWidget;
+import 'reset_password_page_widget.dart' show ResetPasswordPageWidget;
 import 'package:flutter/material.dart';
-import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_animate/flutter_animate.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
 
-class ProductosModel extends FlutterFlowModel<ProductosWidget> {
+class ResetPasswordPageModel extends FlutterFlowModel<ResetPasswordPageWidget> {
   ///  State fields for stateful widgets in this page.
 
   final unfocusNode = FocusNode();
+  // State field(s) for txtEmail widget.
+  FocusNode? txtEmailFocusNode;
+  TextEditingController? txtEmailController;
+  String? Function(BuildContext, String?)? txtEmailControllerValidator;
 
   /// Initialization and disposal methods.
 
@@ -25,6 +23,8 @@ class ProductosModel extends FlutterFlowModel<ProductosWidget> {
 
   void dispose() {
     unfocusNode.dispose();
+    txtEmailFocusNode?.dispose();
+    txtEmailController?.dispose();
   }
 
   /// Action blocks are added here.
