@@ -4,7 +4,6 @@ import '/flutter_flow/flutter_flow_util.dart';
 import '/flutter_flow/flutter_flow_widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
 import 'bs_side_bar_model.dart';
@@ -103,8 +102,8 @@ class _BsSideBarWidgetState extends State<BsSideBarWidget> {
                   text: FFLocalizations.of(context).getText(
                     'syakb4tg' /* Servicios */,
                   ),
-                  icon: FaIcon(
-                    FontAwesomeIcons.shoppingBag,
+                  icon: Icon(
+                    Icons.search,
                     size: 21.0,
                   ),
                   options: FFButtonOptions(
@@ -185,8 +184,8 @@ class _BsSideBarWidgetState extends State<BsSideBarWidget> {
                     '8rdqqyya' /* Formulario */,
                   ),
                   icon: Icon(
-                    Icons.shopping_bag,
-                    size: 21.0,
+                    Icons.newspaper_sharp,
+                    size: 15.0,
                   ),
                   options: FFButtonOptions(
                     width: 150.0,
@@ -219,8 +218,8 @@ class _BsSideBarWidgetState extends State<BsSideBarWidget> {
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
                 FFButtonWidget(
-                  onPressed: () async {
-                    context.pushNamed('CategoriesPage');
+                  onPressed: () {
+                    print('Button pressed ...');
                   },
                   text: FFLocalizations.of(context).getText(
                     'oj58krki' /* Videos */,
@@ -267,6 +266,46 @@ class _BsSideBarWidgetState extends State<BsSideBarWidget> {
                   ),
                   icon: Icon(
                     Icons.person,
+                    size: 15.0,
+                  ),
+                  options: FFButtonOptions(
+                    width: 150.0,
+                    height: 40.0,
+                    padding:
+                        EdgeInsetsDirectional.fromSTEB(24.0, 0.0, 24.0, 0.0),
+                    iconPadding:
+                        EdgeInsetsDirectional.fromSTEB(0.0, 0.0, 0.0, 0.0),
+                    color: Color(0xFFFCDADF),
+                    textStyle: FlutterFlowTheme.of(context).titleSmall.override(
+                          fontFamily: 'Readex Pro',
+                          color: Colors.black,
+                        ),
+                    elevation: 3.0,
+                    borderSide: BorderSide(
+                      color: Colors.transparent,
+                      width: 1.0,
+                    ),
+                    borderRadius: BorderRadius.circular(8.0),
+                  ),
+                ),
+              ],
+            ),
+          ),
+          Padding(
+            padding: EdgeInsetsDirectional.fromSTEB(0.0, 15.0, 0.0, 15.0),
+            child: Row(
+              mainAxisSize: MainAxisSize.max,
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                FFButtonWidget(
+                  onPressed: () async {
+                    context.pushNamed('favorito');
+                  },
+                  text: FFLocalizations.of(context).getText(
+                    'gc5l1v7k' /* Favoritos */,
+                  ),
+                  icon: Icon(
+                    Icons.favorite_border_sharp,
                     size: 15.0,
                   ),
                   options: FFButtonOptions(
