@@ -1,49 +1,66 @@
 // Export pages
-export '/main_content/sobre_nosotros/sobre_nosotros_widget.dart'
+export '/user_u_i/main_content/sobre_nosotros/sobre_nosotros_widget.dart'
     show SobreNosotrosWidget;
 export '/authentication/login_page/login_page_widget.dart' show LoginPageWidget;
 export '/authentication/register/register_widget.dart' show RegisterWidget;
-export '/catalogo/producto1/producto1_widget.dart' show Producto1Widget;
-export '/info_cliente/frm_cliente/frm_cliente_widget.dart'
+export '/user_u_i/catalogo/producto1/producto1_widget.dart'
+    show Producto1Widget;
+export '/user_u_i/info_cliente/frm_cliente/frm_cliente_widget.dart'
     show FrmClienteWidget;
-export '/catalogo/productos/productos_widget.dart' show ProductosWidget;
-export '/perfil/info_perfil/info_perfil_widget.dart' show InfoPerfilWidget;
-export '/info_cliente/editar_cliente/editar_cliente_widget.dart'
+export '/user_u_i/catalogo/productos/productos_widget.dart'
+    show ProductosWidget;
+export '/user_u_i/perfil/info_perfil/info_perfil_widget.dart'
+    show InfoPerfilWidget;
+export '/user_u_i/info_cliente/editar_cliente/editar_cliente_widget.dart'
     show EditarClienteWidget;
-export '/maquillaje/info_maquillaje/info_maquillaje_widget.dart'
+export '/user_u_i/maquillaje/info_maquillaje/info_maquillaje_widget.dart'
     show InfoMaquillajeWidget;
-export '/categoria_maquillaje/cate_maquillaje/cate_maquillaje_widget.dart'
+export '/user_u_i/categoria_maquillaje/cate_maquillaje/cate_maquillaje_widget.dart'
     show CateMaquillajeWidget;
 export '/admin_u_i/products_home/products_home_widget.dart'
     show ProductsHomeWidget;
-export '/main_content/calendar_page/calendar_page_widget.dart'
+export '/user_u_i/main_content/calendar_page/calendar_page_widget.dart'
     show CalendarPageWidget;
-export '/categoria_maquillaje/clases_automaquillaje/clases_automaquillaje_widget.dart'
+export '/user_u_i/categoria_maquillaje/clases_automaquillaje/clases_automaquillaje_widget.dart'
     show ClasesAutomaquillajeWidget;
-export '/categoria_maquillaje/asesoria/asesoria_widget.dart'
+export '/user_u_i/categoria_maquillaje/asesoria/asesoria_widget.dart'
     show AsesoriaWidget;
-export '/categoria_maquillaje/maquilaje_eventos/maquilaje_eventos_widget.dart'
+export '/user_u_i/categoria_maquillaje/maquilaje_eventos/maquilaje_eventos_widget.dart'
     show MaquilajeEventosWidget;
-export '/categoria_maquillaje/paquetes_novia/paquetes_novia_widget.dart'
+export '/user_u_i/categoria_maquillaje/paquetes_novia/paquetes_novia_widget.dart'
     show PaquetesNoviaWidget;
-export '/categoria_maquillaje/skin_care/skin_care_widget.dart'
+export '/user_u_i/categoria_maquillaje/skin_care/skin_care_widget.dart'
     show SkinCareWidget;
-export '/categoria_maquillaje/venta_maquillaje/venta_maquillaje_widget.dart'
+export '/user_u_i/categoria_maquillaje/venta_maquillaje/venta_maquillaje_widget.dart'
     show VentaMaquillajeWidget;
-export '/catalogo/producto2/producto2_widget.dart' show Producto2Widget;
-export '/catalogo/producto3/producto3_widget.dart' show Producto3Widget;
-export '/catalogo/producto4/producto4_widget.dart' show Producto4Widget;
-export '/catalogo/producto5/producto5_widget.dart' show Producto5Widget;
-export '/catalogo/producto6/producto6_widget.dart' show Producto6Widget;
-export '/catalogo/producto7/producto7_widget.dart' show Producto7Widget;
-export '/catalogo/producto8/producto8_widget.dart' show Producto8Widget;
-export '/catalogo/producto9/producto9_widget.dart' show Producto9Widget;
-export '/catalogo/producto10/producto10_widget.dart' show Producto10Widget;
-export '/catalogo/producto11/producto11_widget.dart' show Producto11Widget;
+export '/user_u_i/catalogo/producto2/producto2_widget.dart'
+    show Producto2Widget;
+export '/user_u_i/catalogo/producto3/producto3_widget.dart'
+    show Producto3Widget;
+export '/user_u_i/catalogo/producto4/producto4_widget.dart'
+    show Producto4Widget;
+export '/user_u_i/catalogo/producto5/producto5_widget.dart'
+    show Producto5Widget;
+export '/user_u_i/catalogo/producto6/producto6_widget.dart'
+    show Producto6Widget;
+export '/user_u_i/catalogo/producto7/producto7_widget.dart'
+    show Producto7Widget;
+export '/user_u_i/catalogo/producto8/producto8_widget.dart'
+    show Producto8Widget;
+export '/user_u_i/catalogo/producto9/producto9_widget.dart'
+    show Producto9Widget;
+export '/user_u_i/catalogo/producto10/producto10_widget.dart'
+    show Producto10Widget;
+export '/user_u_i/catalogo/producto11/producto11_widget.dart'
+    show Producto11Widget;
 export '/admin_u_i/categories_page/categories_page_widget.dart'
     show CategoriesPageWidget;
 export '/admin_u_i/brands_page/brands_page_widget.dart' show BrandsPageWidget;
-export '/catalogo/producto1_copy/producto1_copy_widget.dart'
-    show Producto1CopyWidget;
+export '/user_u_i/catalogo/info_productos/info_productos_widget.dart'
+    show InfoProductosWidget;
 export '/admin_u_i/appointment_page/appointment_page_widget.dart'
     show AppointmentPageWidget;
+export '/user_u_i/catalogo/favorito/favorito_widget.dart' show FavoritoWidget;
+export '/admin_u_i/users_page/users_page_widget.dart' show UsersPageWidget;
+export '/authentication/reset_password_page/reset_password_page_widget.dart'
+    show ResetPasswordPageWidget;
