@@ -1,16 +1,20 @@
+import '/auth/firebase_auth/auth_util.dart';
 import '/backend/backend.dart';
+import '/componentes/bs_side_bar/bs_side_bar_widget.dart';
+import '/flutter_flow/flutter_flow_animations.dart';
 import '/flutter_flow/flutter_flow_icon_button.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
 import '/flutter_flow/flutter_flow_widgets.dart';
-import 'dart:ui';
-import 'producto1_copy_widget.dart' show Producto1CopyWidget;
+import 'favorito_widget.dart' show FavoritoWidget;
 import 'package:flutter/material.dart';
+import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter_animate/flutter_animate.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
 
-class Producto1CopyModel extends FlutterFlowModel<Producto1CopyWidget> {
+class FavoritoModel extends FlutterFlowModel<FavoritoWidget> {
   ///  State fields for stateful widgets in this page.
 
   final unfocusNode = FocusNode();
