@@ -2,12 +2,15 @@ import '/auth/firebase_auth/auth_util.dart';
 import '/backend/backend.dart';
 import '/backend/firebase_storage/storage.dart';
 import '/componentes/bs_side_bar/bs_side_bar_widget.dart';
+import '/flutter_flow/flutter_flow_drop_down.dart';
 import '/flutter_flow/flutter_flow_icon_button.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
 import '/flutter_flow/flutter_flow_widgets.dart';
+import '/flutter_flow/form_field_controller.dart';
 import '/flutter_flow/upload_data.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:google_fonts/google_fonts.dart';
@@ -40,23 +43,9 @@ class _InfoPerfilWidgetState extends State<InfoPerfilWidget> {
         TextEditingController(text: currentUserEmail);
     _model.txtCorreoFocusNode ??= FocusNode();
 
-    _model.txNacimientoController ??= TextEditingController();
-    _model.txNacimientoFocusNode ??= FocusNode();
-
     _model.txtTelefonoController ??=
         TextEditingController(text: currentPhoneNumber);
     _model.txtTelefonoFocusNode ??= FocusNode();
-
-    _model.txtContrasenaController ??= TextEditingController();
-    _model.txtContrasenaFocusNode ??= FocusNode();
-
-    WidgetsBinding.instance.addPostFrameCallback((_) => setState(() {
-          _model.txNacimientoController?.text = dateTimeFormat(
-            'd/M/y',
-            currentUserDocument?.birthday,
-            locale: FFLocalizations.of(context).languageCode,
-          );
-        }));
   }
 
   @override
@@ -123,7 +112,48 @@ class _InfoPerfilWidgetState extends State<InfoPerfilWidget> {
                 },
               ),
             ),
-            actions: [],
+            actions: [
+              FlutterFlowIconButton(
+                borderColor: Colors.transparent,
+                borderRadius: 20.0,
+                borderWidth: 1.0,
+                buttonSize: 40.0,
+                icon: Icon(
+                  Icons.logout_rounded,
+                  color: Colors.white,
+                  size: 24.0,
+                ),
+                onPressed: () async {
+                  var confirmDialogResponse = await showDialog<bool>(
+                        context: context,
+                        builder: (alertDialogContext) {
+                          return AlertDialog(
+                            title: Text('Cerrar Sesión'),
+                            content: Text('¿Está seguro/a qué desea salir?'),
+                            actions: [
+                              TextButton(
+                                onPressed: () =>
+                                    Navigator.pop(alertDialogContext, false),
+                                child: Text('Cancelar'),
+                              ),
+                              TextButton(
+                                onPressed: () =>
+                                    Navigator.pop(alertDialogContext, true),
+                                child: Text('Confirmar'),
+                              ),
+                            ],
+                          );
+                        },
+                      ) ??
+                      false;
+                  GoRouter.of(context).prepareAuthEvent();
+                  await authManager.signOut();
+                  GoRouter.of(context).clearRedirectLocation();
+
+                  context.goNamedAuth('LoginPage', context.mounted);
+                },
+              ),
+            ],
             centerTitle: false,
             toolbarHeight: 70.0,
             elevation: 2.0,
@@ -408,76 +438,6 @@ class _InfoPerfilWidgetState extends State<InfoPerfilWidget> {
                               23.0, 30.0, 23.0, 0.0),
                           child: AuthUserStreamWidget(
                             builder: (context) => TextFormField(
-                              controller: _model.txNacimientoController,
-                              focusNode: _model.txNacimientoFocusNode,
-                              autofocus: true,
-                              textCapitalization: TextCapitalization.none,
-                              obscureText: false,
-                              decoration: InputDecoration(
-                                labelText: FFLocalizations.of(context).getText(
-                                  '82hi21ts' /* Fecha de nacimiento */,
-                                ),
-                                labelStyle: FlutterFlowTheme.of(context)
-                                    .labelMedium
-                                    .override(
-                                      fontFamily: 'Readex Pro',
-                                      color: Color(0xFF14181B),
-                                    ),
-                                hintStyle:
-                                    FlutterFlowTheme.of(context).labelMedium,
-                                enabledBorder: OutlineInputBorder(
-                                  borderSide: BorderSide(
-                                    color:
-                                        FlutterFlowTheme.of(context).alternate,
-                                    width: 2.0,
-                                  ),
-                                  borderRadius: BorderRadius.circular(20.0),
-                                ),
-                                focusedBorder: OutlineInputBorder(
-                                  borderSide: BorderSide(
-                                    color: Color(0xFFF9F9F9),
-                                    width: 2.0,
-                                  ),
-                                  borderRadius: BorderRadius.circular(20.0),
-                                ),
-                                errorBorder: OutlineInputBorder(
-                                  borderSide: BorderSide(
-                                    color: FlutterFlowTheme.of(context).error,
-                                    width: 2.0,
-                                  ),
-                                  borderRadius: BorderRadius.circular(20.0),
-                                ),
-                                focusedErrorBorder: OutlineInputBorder(
-                                  borderSide: BorderSide(
-                                    color: FlutterFlowTheme.of(context).error,
-                                    width: 2.0,
-                                  ),
-                                  borderRadius: BorderRadius.circular(20.0),
-                                ),
-                                filled: true,
-                                fillColor: Color(0xFFF9F9F9),
-                                contentPadding: EdgeInsetsDirectional.fromSTEB(
-                                    20.0, 20.0, 20.0, 20.0),
-                              ),
-                              style: FlutterFlowTheme.of(context).bodyMedium,
-                              keyboardType: TextInputType.datetime,
-                              validator: _model.txNacimientoControllerValidator
-                                  .asValidator(context),
-                            ),
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                  Row(
-                    mainAxisSize: MainAxisSize.max,
-                    children: [
-                      Expanded(
-                        child: Padding(
-                          padding: EdgeInsetsDirectional.fromSTEB(
-                              23.0, 30.0, 23.0, 0.0),
-                          child: AuthUserStreamWidget(
-                            builder: (context) => TextFormField(
                               controller: _model.txtTelefonoController,
                               focusNode: _model.txtTelefonoFocusNode,
                               autofocus: true,
@@ -538,95 +498,137 @@ class _InfoPerfilWidgetState extends State<InfoPerfilWidget> {
                       ),
                     ],
                   ),
-                  Row(
-                    mainAxisSize: MainAxisSize.max,
-                    children: [
-                      Expanded(
-                        child: Padding(
-                          padding: EdgeInsetsDirectional.fromSTEB(
-                              23.0, 30.0, 23.0, 0.0),
-                          child: TextFormField(
-                            controller: _model.txtContrasenaController,
-                            focusNode: _model.txtContrasenaFocusNode,
-                            autofocus: true,
-                            obscureText: !_model.txtContrasenaVisibility,
-                            decoration: InputDecoration(
-                              labelText: FFLocalizations.of(context).getText(
-                                'itpbs5cf' /* Digite su contraseña */,
-                              ),
-                              labelStyle: FlutterFlowTheme.of(context)
-                                  .labelMedium
-                                  .override(
-                                    fontFamily: 'Readex Pro',
-                                    color: Color(0xFF14181B),
-                                  ),
-                              hintStyle:
-                                  FlutterFlowTheme.of(context).labelMedium,
-                              enabledBorder: OutlineInputBorder(
-                                borderSide: BorderSide(
-                                  color: FlutterFlowTheme.of(context).alternate,
-                                  width: 2.0,
-                                ),
-                                borderRadius: BorderRadius.circular(20.0),
-                              ),
-                              focusedBorder: OutlineInputBorder(
-                                borderSide: BorderSide(
-                                  color: Color(0xFFF9F9F9),
-                                  width: 2.0,
-                                ),
-                                borderRadius: BorderRadius.circular(20.0),
-                              ),
-                              errorBorder: OutlineInputBorder(
-                                borderSide: BorderSide(
-                                  color: FlutterFlowTheme.of(context).error,
-                                  width: 2.0,
-                                ),
-                                borderRadius: BorderRadius.circular(20.0),
-                              ),
-                              focusedErrorBorder: OutlineInputBorder(
-                                borderSide: BorderSide(
-                                  color: FlutterFlowTheme.of(context).error,
-                                  width: 2.0,
-                                ),
-                                borderRadius: BorderRadius.circular(20.0),
-                              ),
-                              filled: true,
-                              fillColor: Color(0xFFF9F9F9),
-                              contentPadding: EdgeInsetsDirectional.fromSTEB(
-                                  20.0, 20.0, 20.0, 20.0),
-                              suffixIcon: InkWell(
-                                onTap: () => setState(
-                                  () => _model.txtContrasenaVisibility =
-                                      !_model.txtContrasenaVisibility,
-                                ),
-                                focusNode: FocusNode(skipTraversal: true),
-                                child: Icon(
-                                  _model.txtContrasenaVisibility
-                                      ? Icons.visibility_outlined
-                                      : Icons.visibility_off_outlined,
-                                  size: 22,
-                                ),
-                              ),
-                            ),
-                            style: FlutterFlowTheme.of(context).bodyMedium,
-                            validator: _model.txtContrasenaControllerValidator
-                                .asValidator(context),
+                  Padding(
+                    padding:
+                        EdgeInsetsDirectional.fromSTEB(0.0, 16.0, 0.0, 0.0),
+                    child: AuthUserStreamWidget(
+                      builder: (context) => FFButtonWidget(
+                        onPressed: () async {
+                          final _datePickedDate = await showDatePicker(
+                            context: context,
+                            initialDate: (_model.datePicked ?? DateTime.now()),
+                            firstDate: DateTime(1900),
+                            lastDate: (_model.datePicked ?? DateTime.now()),
+                          );
+
+                          if (_datePickedDate != null) {
+                            safeSetState(() {
+                              _model.datePicked = DateTime(
+                                _datePickedDate.year,
+                                _datePickedDate.month,
+                                _datePickedDate.day,
+                              );
+                            });
+                          }
+                        },
+                        text: valueOrDefault<String>(
+                          dateTimeFormat(
+                            'd/M/y',
+                            currentUserDocument?.birthday,
+                            locale: FFLocalizations.of(context).languageCode,
                           ),
+                          'Fecha de Nacimiento',
+                        ),
+                        options: FFButtonOptions(
+                          width: 340.0,
+                          height: 40.0,
+                          padding: EdgeInsetsDirectional.fromSTEB(
+                              24.0, 0.0, 24.0, 0.0),
+                          iconPadding: EdgeInsetsDirectional.fromSTEB(
+                              0.0, 0.0, 0.0, 0.0),
+                          color: Color(0xFFFF1493),
+                          textStyle:
+                              FlutterFlowTheme.of(context).titleSmall.override(
+                                    fontFamily: 'Readex Pro',
+                                    color: Colors.white,
+                                  ),
+                          elevation: 3.0,
+                          borderSide: BorderSide(
+                            color: Colors.transparent,
+                            width: 1.0,
+                          ),
+                          borderRadius: BorderRadius.circular(8.0),
                         ),
                       ),
-                    ],
+                    ),
                   ),
                   Padding(
                     padding:
-                        EdgeInsetsDirectional.fromSTEB(0.0, 30.0, 0.0, 0.0),
+                        EdgeInsetsDirectional.fromSTEB(0.0, 20.0, 0.0, 0.0),
+                    child: AuthUserStreamWidget(
+                      builder: (context) =>
+                          StreamBuilder<List<SkinTypesRecord>>(
+                        stream: querySkinTypesRecord(
+                          queryBuilder: (skinTypesRecord) =>
+                              skinTypesRecord.orderBy('skin_type'),
+                        ),
+                        builder: (context, snapshot) {
+                          // Customize what your widget looks like when it's loading.
+                          if (!snapshot.hasData) {
+                            return Center(
+                              child: SizedBox(
+                                width: 50.0,
+                                height: 50.0,
+                                child: CircularProgressIndicator(
+                                  valueColor: AlwaysStoppedAnimation<Color>(
+                                    FlutterFlowTheme.of(context).primary,
+                                  ),
+                                ),
+                              ),
+                            );
+                          }
+                          List<SkinTypesRecord> ddSkinSkinTypesRecordList =
+                              snapshot.data!;
+                          return FlutterFlowDropDown<String>(
+                            controller: _model.ddSkinValueController ??=
+                                FormFieldController<String>(
+                              _model.ddSkinValue ??= valueOrDefault(
+                                  currentUserDocument?.skinType, ''),
+                            ),
+                            options: ddSkinSkinTypesRecordList
+                                .map((e) => e.skinType)
+                                .toList(),
+                            onChanged: (val) =>
+                                setState(() => _model.ddSkinValue = val),
+                            width: 340.0,
+                            height: 50.0,
+                            textStyle: FlutterFlowTheme.of(context).bodyMedium,
+                            hintText: FFLocalizations.of(context).getText(
+                              'wfugbnwo' /* Tipo de piel */,
+                            ),
+                            icon: Icon(
+                              Icons.keyboard_arrow_down_rounded,
+                              color: FlutterFlowTheme.of(context).secondaryText,
+                              size: 24.0,
+                            ),
+                            fillColor: FlutterFlowTheme.of(context)
+                                .secondaryBackground,
+                            elevation: 2.0,
+                            borderColor: FlutterFlowTheme.of(context).alternate,
+                            borderWidth: 2.0,
+                            borderRadius: 8.0,
+                            margin: EdgeInsetsDirectional.fromSTEB(
+                                16.0, 4.0, 16.0, 4.0),
+                            hidesUnderline: true,
+                            isSearchable: false,
+                            isMultiSelect: false,
+                          );
+                        },
+                      ),
+                    ),
+                  ),
+                  Padding(
+                    padding:
+                        EdgeInsetsDirectional.fromSTEB(0.0, 20.0, 0.0, 0.0),
                     child: FFButtonWidget(
                       onPressed: () async {
                         await currentUserReference!
                             .update(createUsersRecordData(
                           email: _model.txtCorreoController.text,
                           displayName: _model.txtNombreController.text,
-                          photoUrl: '',
                           phoneNumber: _model.txtTelefonoController.text,
+                          birthday: _model.datePicked,
+                          skinType: _model.ddSkinValue,
                         ));
                       },
                       text: FFLocalizations.of(context).getText(
